fix custom lints
